ARRISXI6-1897: DSMgr crash if AVR in the middle
Reason for change:  size_t is unsigned and loop never breaks.
Test Procedure:  TV with 4K and AVR without 4K support
Risks: No

(cherry picked from commit f2c2a9b20d852cb533cfa21d2f702d18405fba62)

+++ dsmgr/dsMgr.c
@@ -595,8 +595,8 @@ static int  _SetResolution(int* handle,dsVideoPortType_t PortType)
                 		for (i = numResolutions-1; i >= 0; i--)
                 		{
                         		setResn = &(edidData.suppResolutionList[i]);
-                        		size_t pNumResolutions = dsUTL_DIM(kResolutions);
-                        		for (size_t j = pNumResolutions-1; j >=0; j--)
+                        		int pNumResolutions = dsUTL_DIM(kResolutions);
+                        		for (int j = pNumResolutions-1; j >=0; j--)
 		        		{
 		                		dsVideoPortResolution_t *pfResolution = &kResolutions[j];
 		                		if (0 == (strcmp(pfResolution->name,setResn->name)))
